+++ srcs/ast/ArrayElementNode.h
@@ -7,7 +7,7 @@
 class ArrayElementNode : public Node
 {
 public:
-  static ArrayElementNode* Create(const char* _value, int n);
+  static ArrayElementNode* Create(const char* _value, double n);
 	static ArrayElementNode* Create(const char* _value1, const char* _value2);
 	std::string getStr1() { return value1; }
 	std::string getStr2() { return value2; }
@@ -16,7 +16,7 @@ public:
 	int getArrEleType();
 
 protected:
-	ArrayElementNode(const char* _value, int n);
+	ArrayElementNode(const char* _value, double n);
 	ArrayElementNode(const char* _value1, const char* _value2);
 	std::string value1;
 	std::string value2;
